refactor: replace matrix usage with ReleaseMatrix for build job configurations
