+++ crates/forge_ci/src/workflows/release_homebrew.rs
@@ -2,11 +2,11 @@ use gh_workflow_tailcall::generate::Generate;
 use gh_workflow_tailcall::*;
 
 use crate::jobs::{create_build_release_job_for_publishing, create_homebrew_release_job};
-use crate::matrix;
+use crate::release_matrix::{self, ReleaseMatrix};
 
 /// Generate homebrew release workflow
 pub fn generate_homebrew_workflow() {
-    let build_job = create_build_release_job_for_publishing(matrix::create_matrix());
+    let build_job = create_build_release_job_for_publishing(ReleaseMatrix::default());
     let homebrew_release_job = create_homebrew_release_job().add_needs(build_job.clone());
     let homebrew_workflow = Workflow::default()
         .name("Homebrew Release")
